web: Move some polyfill code out of RufflePlayer
