+++ web/packages/core/src/ruffle-embed.ts
@@ -1,4 +1,4 @@
-import { RufflePlayer } from "./ruffle-player";
+import { copyElement, RufflePlayer } from "./ruffle-player";
 import {
     getPolyfillOptions,
     isFallbackElement,
@@ -139,8 +139,68 @@ export class RuffleEmbed extends RufflePlayer {
     static fromNativeEmbedElement(elem: Element): RuffleEmbed {
         const externalName = registerElement("ruffle-embed", RuffleEmbed);
         const ruffleObj = document.createElement(externalName) as RuffleEmbed;
-        ruffleObj.copyElement(elem);
+        copyElement(elem, ruffleObj);
 
         return ruffleObj;
+    }
+
+    /**
+     * Polyfill of height getter
+     *
+     * @ignore
+     * @internal
+     */
+    get height(): string {
+        return this.getAttribute("height") || "";
+    }
+
+    /**
+     * Polyfill of height setter
+     *
+     * @ignore
+     * @internal
+     */
+    set height(height: string) {
+        this.setAttribute("height", height);
+    }
+
+    /**
+     * Polyfill of width getter
+     *
+     * @ignore
+     * @internal
+     */
+    get width(): string {
+        return this.getAttribute("width") || "";
+    }
+
+    /**
+     * Polyfill of width setter
+     *
+     * @ignore
+     * @internal
+     */
+    set width(widthVal: string) {
+        this.setAttribute("width", widthVal);
+    }
+
+    /**
+     * Polyfill of type getter
+     *
+     * @ignore
+     * @internal
+     */
+    get type(): string {
+        return this.getAttribute("type") || "";
+    }
+
+    /**
+     * Polyfill of type setter
+     *
+     * @ignore
+     * @internal
+     */
+    set type(typeVal: string) {
+        this.setAttribute("type", typeVal);
     }
 }
